requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
Pillow==7.2.0
pkg-resources==0.0.0
PyAudio==0.2.11
pydub==0.24.1
PyPDF2==1.26.0
python-docx==0.8.10
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
urllib3==1.25.10
Werkzeug==1.0.1
wikipedia==1.4.0
